requirements.txt
boto3
opencv-python
python-dotenv
loguru
picamera2
libcamera
